refactor: refactor tests folder
start #48
